stm32h7: move startup hooks to board directories for easier customization
